requirements.txt
asgiref==3.7.2
Django==3.2.16
django-bootstrap5==22.2
django-debug-toolbar==3.8.1
Pillow==9.3.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
